ASSETS-47371: more testing
