+++ .github/actions/lint-test-coverage/action.yaml
@@ -18,10 +18,6 @@ runs:
       run: npm run lint
       shell: bash
 
-    - name: Test
-      run: npm run test
-      shell: bash
-
     - name: Upload coverage to Codecov
       if: ${{ inputs.upload_coverage == 'true' }}
       uses: codecov/codecov-action@v5
